requirements.txt
Django==1.9.2
django-allauth==0.24.1
django-rest-auth==0.6.0
djangorestframework==3.3.2
drf-nested-routers==0.11.1
oauthlib==1.0.3
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
